make check 修正 issue #2 の解決、軽微なコンパイラ警告の修正

+++ libtumbler/test/speaker_test.cpp
@@ -55,13 +55,13 @@ int main(int argc, char** argv)
     using namespace tumbler;
 	std::vector<short> audio = FileReader("data/test1.raw");
 	std::vector<short> ad;
-	for(int i=0;i<audio.size();++i){
+	for(size_t i=0;i<audio.size();++i){
 		ad.push_back(audio[i]);
 	}
-	for(int i=0;i<audio.size();++i){
+	for(size_t i=0;i<audio.size();++i){
 		ad.push_back(audio[i]);
 	}
-	for(int i=0;i<audio.size();++i){
+	for(size_t i=0;i<audio.size();++i){
 		ad.push_back(audio[i]);
 	}
     {
